extractor/requirements.txt
Pillow
pyocr
Wand
